refactor(livechat): Rearrange embeddable script imports/exports
