+++ packages/livechat/src/lib/room.js
@@ -8,8 +8,8 @@ import { upsert } from '../helpers/upsert';
 import { store, initialState } from '../store';
 import { normalizeAgent } from './api';
 import Commands from './commands';
+import { parentCall } from './iframe';
 import { loadConfig, processUnread } from './main';
-import { parentCall } from './parentCall';
 import { createToken } from './random';
 import { normalizeMessage, normalizeMessages } from './threads';
 import { handleTranscript } from './transcript';
