+++ packages/livechat/src/routes/Chat/container.js
@@ -20,8 +20,8 @@ import { useRoomMessagesSubscription } from '../../hooks/useRoomMessagesSubscrip
 import { useUserActivitySubscription } from '../../hooks/useUserActivitySubscription';
 import { normalizeQueueAlert } from '../../lib/api';
 import constants from '../../lib/constants';
+import { parentCall, runCallbackEventEmitter } from '../../lib/iframe';
 import { getLastReadMessage, loadConfig, processUnread, shouldMarkAsUnread } from '../../lib/main';
-import { parentCall, runCallbackEventEmitter } from '../../lib/parentCall';
 import { createToken } from '../../lib/random';
 import { initRoom, loadMessages, loadMoreMessages, defaultRoomParams, getGreetingMessages } from '../../lib/room';
 import store from '../../store';
