+++ packages/livechat/src/routes/LeaveMessage/index.tsx
@@ -18,7 +18,7 @@ import { createClassName } from '../../helpers/createClassName';
 import { parseOfflineMessage } from '../../helpers/parseOfflineMessage';
 import { sortArrayByColumn } from '../../helpers/sortArrayByColumn';
 import { validateEmail } from '../../lib/email';
-import { parentCall } from '../../lib/parentCall';
+import { parentCall } from '../../lib/iframe';
 import { createToken } from '../../lib/random';
 import { StoreContext } from '../../store';
