+++ packages/livechat/src/routes/TriggerMessage/component.tsx
@@ -7,7 +7,7 @@ import styles from './styles.scss';
 import Screen from '../../components/Screen';
 import type { ScreenContextValue } from '../../components/Screen/ScreenProvider';
 import { createClassName } from '../../helpers/createClassName';
-import { parentCall } from '../../lib/parentCall';
+import { parentCall } from '../../lib/iframe';
 import type { StoreState } from '../../store';
 
 type TriggerMessageProps = {
