+++ packages/livechat/src/routes/TriggerMessage/container.tsx
@@ -6,7 +6,7 @@ import TriggerMessage from './component';
 import { ScreenContext } from '../../components/Screen/ScreenProvider';
 import { canRenderMessage } from '../../helpers/canRenderMessage';
 import { formatAgent } from '../../helpers/formatAgent';
-import { parentCall } from '../../lib/parentCall';
+import { parentCall } from '../../lib/iframe';
 import { StoreContext } from '../../store';
 
 export const TriggerMessageContainer: FunctionalComponent<{ path: string }> = ({ ref }) => {
